+++ deerlab/correctphase.py
@@ -2,7 +2,7 @@ import numpy as np
 import scipy.optimize as opt
 from deerlab.utils import isempty
 
-def correctphase(V, phase = [], imagoffset=False, full_output=False):
+def correctphase(V, phase='posrealint', full_output=False):
 # ==========================================================================
     r"""
     Phase correction of complex-valued data
@@ -10,8 +10,6 @@ def correctphase(V, phase = [], imagoffset=False, full_output=False):
     Performs a phase optimization on the complex-valued data ``V`` by determining a phase
     rotation of ``V`` that minimizes the imaginary component.
     
-    Also, the phase can be corrected manually by specifying  a phase ``phase``, in radians.
-
     Two-dimensional datasets ``V2D``, e.g. from multiple scans measurements, can be provided, 
     and the phase correction will be done on each trace individually. The first dimension ``V2D[:,i]``
     must contain the single traces. An array of phases ``phases`` can be specified to manually correct the traces.
@@ -20,10 +18,16 @@ def correctphase(V, phase = [], imagoffset=False, full_output=False):
     ----------
     V : array_like or list of array_like
         Complex-valued signals or list thereof.
-    phase : float scalar, optional
-        Phase shift for manual correction, in radians. 
-    imagoffset : boolean, optional
-        Enables/Disables the fitting and correction of an imaginary offset, by default disabled.
+
+    phase : string, optional
+        Criterion for selection of correction phase. 
+
+        * ``'posrealint'`` - Select the phase that gives the largest positive integral of the real part.
+        * ``'negrealint'`` - Select the phase that gives the largest negative integral of the real part.
+        * ``'close'`` - Select the phase closest to the average phase of the original data.
+
+        The default behaviour is ``'posrealint'``.
+
     full_output : boolean, optional
         If enabled, the function will return additional output arguments, by default disabled.
 
@@ -31,94 +35,61 @@ def correctphase(V, phase = [], imagoffset=False, full_output=False):
     -------
     Vr : ndarray
         Real part of the phase corrected dataset.
+
     Vi : ndarray (if full_output==True)
         Imaginary part of the phase corrected dataset.
+
     phase : float scalar (if full_output==True)
         Fitted phase used for correction, in radians.    
-    imoffset : float scalar (if full_output==True)
-        Fitted imaginary offset used for correction.
 
     """
 
-    if V.ndim == 1:
+    V_2d = V.copy()
+    if V_2d.ndim == 1:
         Ntraces = 1
-        V = V[:,np.newaxis]
+        V_2d = V_2d[:,np.newaxis]
     else:
-        Ntraces = V.shape[1]
-    n = V.shape[0]
+        Ntraces = V_2d.shape[1]
+    n = V_2d.shape[0]
+ 
+    phaseopt = np.zeros(Ntraces)
+    for i in range(Ntraces):
+        V_ = V_2d[:,i]
 
-    # Determine if phase must be fitted or has been passed
-    if isempty(phase):
-        fitPhase = True
-    else: 
-        fitPhase = False
-        phase = np.atleast_1d(phase)
-        if len(phase) != Ntraces:
-            raise ValueError('The number of input phases must agree with the number of traces.') 
+        # Get the average phase in the signal
+        phaseav = np.mean(np.angle(V_))
 
-    # Phase/offset fitting
-    #-------------------------------------------------------------------------------
-    ImagOffset = np.zeros(Ntraces)
-    if fitPhase:
-        phase = np.zeros(Ntraces)
-        for i in range(Ntraces):
-            par0 = []
-            FitRange = np.arange(round(n/8),n) # use only last 7/8 of data for phase/offset correction
-            V_ = V[FitRange,i]
-            par0.append(np.mean(np.angle(V_))) # use average phase as initial value
-            if imagoffset:
-                par0.append(np.mean(np.imag(V_))) # use average offset as initial value
-            fun = lambda par: _imaginarynorm(par,V_)
+        # Objective function for optimization of the phase
+        phase_objfcn = lambda phase: np.linalg.norm(np.imag(V_*np.exp(-1j*phase)))
 
-            pars = opt.fmin(fun,par0,maxfun=1e5,maxiter=1e5,disp=False)
-            phase[i] = pars[0]
-            if imagoffset:
-                ImagOffset[i] = pars[1]
-    else:
-        if imagoffset:
-            # Fit only imaginary offset        
-            for i in range(Ntraces):
-                par0 = 0
-                fun = lambda offset: _imaginarynorm([phase, offset],V[:,i])
-                ImagOffset[i] = opt.fmin(fun,par0)
+        pars = opt.fmin(phase_objfcn,phaseav,maxfun=1e5,maxiter=1e5,disp=False)
+        
+        # Get the two phases that minimize the imaginary part
+        phases = [pars[0], pars[0]+np.pi]
+        realint = [np.sum(V_*np.exp(-1j*phase)) for phase in phases]
+            
+        if phase == 'posrealint':
+            phaseopt[i] =  phases[np.argmax(realint)]
+        elif phase == 'negrealint':
+            phaseopt[i] =  phases[np.argmin(realint)]
+        elif phase == 'close':
+            phaseopt[i] =  phases[np.argmin(abs(phases - phaseav))]
 
-    ImagOffset = ImagOffset*1j
-
-    # Apply phase/offset correction
-    ph = np.exp(1j*phase)
-    Vc = (V - ImagOffset)/ph
+        # Apply phase
+        V_2d[:,i] = V_*np.exp(-1j*phaseopt[i])
 
     if Ntraces==1:
-        Vc = np.squeeze(Vc)
+        V_2d = np.squeeze(V_2d)
 
     # Output
-    Vreal = np.real(Vc)
-    Vimag = np.imag(Vc)
-    phase = np.angle(ph) # map phase angle to [-pi,pi) interval
+    Vreal = np.real(V_2d)
+    Vimag = np.imag(V_2d)
+    # Map phase angle to [-pi,pi) interval
+    phase = np.angle(np.exp(-1j*phaseopt)) 
 
     if full_output:
-        return Vreal,Vimag,phase,ImagOffset
+        return Vreal,Vimag,phase
     else:
         return Vreal
 
-# ==========================================================================
-
-
-def _imaginarynorm(params,V):
-# ==========================================================================
-    """
-    Computes norm of the imaginary part of phase-corrected data from zero before
-    phase correction, an offset can be subtracted from the imaginary part.
-    """
-
-    phase = params[0]
-    if len(params)>1:
-        imoffsets = params[1]
-    else:
-        imoffsets = 0
-
-    Vcorr = (V-1j*imoffsets)*np.exp(-1j*phase)
-    ImagNorm = np.linalg.norm(np.imag(Vcorr))
-
-    return ImagNorm
 # ==========================================================================