+++ CHANGELOG.md
@@ -55,7 +55,11 @@ Release v0.13.0 - March 2021
     - Added internal validation step to avoid non-sensical results when confounding the syntax ([#91](https://github.com/JeschkeLab/DeerLab/pull/91)).
 - ``regparamrange``:
     - Implemented new CSD algorithm to avoid LAPACK library crashes encountered when using multiple DeerLab functions calling ``regparamrange`` internally ([#68](https://github.com/JeschkeLab/DeerLab/pull/68)).
-
+- ``correctphase``: 
+    - Implement new keyword ``phase`` to select the criterion for optimizing the phase for correction ([#114](https://github.com/JeschkeLab/DeerLab/issues/114), [#131](https://github.com/JeschkeLab/DeerLab/pull0/131)).
+    - Deprecated imaginary offset fitting ([#131](https://github.com/JeschkeLab/DeerLab/pull0/131)). 
+    - Deprecated manual phase correction. Manual correction can be done by the user and is now described in the beginner's guide ([#131](https://github.com/JeschkeLab/DeerLab/pull0/131)). 
+    
 -------------------------------
 
 Release v0.12.2 - October 2020